refactor: move use heder hook to hooks folder
