+++ eslint.config.js
@@ -1,4 +1,4 @@
-{
+export default [{
   "root": true,
   "extends": ["react-app", "plugin:prettier/recommended", "./.eslintrc.base"]
-}
+}]
