+++ packages/widget/src/pages/SendToWallet/SendToConfiguredWalletPage.tsx
@@ -16,7 +16,7 @@ import { useToAddressRequirements } from '../../hooks/useToAddressRequirements.j
 import { useWidgetConfig } from '../../providers/WidgetProvider/WidgetProvider.js';
 import { useBookmarkActions } from '../../stores/bookmarks/useBookmarkActions.js';
 import { useFieldActions } from '../../stores/form/useFieldActions.js';
-import { useHeader } from '../../stores/header/useHeaderStore.js';
+import { useHeader } from '../../hooks/useHeader.js';
 import type { ToAddress } from '../../types/widget.js';
 import { defaultChainIdsByType } from '../../utils/chainType.js';
 import { navigationRoutes } from '../../utils/navigationRoutes.js';
