+++ packages/widget/src/pages/SettingsPage/SettingsPage.tsx
@@ -8,7 +8,7 @@ import { SettingsList } from './SettingsCard/SettingCard.style.js';
 import { SettingsCardAccordion } from './SettingsCard/SettingsAccordian.js';
 import { SlippageSettings } from './SlippageSettings/SlippageSettings.js';
 import { ThemeSettings } from './ThemeSettings.js';
-import { useHeader } from '../../stores/header/useHeaderStore.js';
+import { useHeader } from '../../hooks/useHeader.js';
 import { useTranslation } from 'react-i18next';
 
 export const SettingsPage = () => {
